Agrego opción de shuffle labels y precomputed model configurations

+++ shared/train_classifier.py
@@ -5,10 +5,7 @@ import math
 
 from sklearn.model_selection import StratifiedKFold, KFold
 from sklearn.linear_model import LogisticRegression as LR
-from sklearn.ensemble import RandomForestClassifier
-from sklearn.discriminant_analysis import LinearDiscriminantAnalysis as LDA
 from sklearn.svm import SVC
-from sklearn.naive_bayes import GaussianNB as NB
 from sklearn.neighbors import KNeighborsClassifier as KNN
 from sklearn.model_selection import train_test_split,StratifiedShuffleSplit,LeavePOut
 from xgboost import XGBClassifier as xgboost
@@ -29,44 +26,14 @@ from utils import *
 from expected_cost.ec import *
 from expected_cost.utils import *
 
-parallel = True
-
-l2ocv = False
-
-test_size = 0
-
+##---------------------------------PARAMETERS---------------------------------##
 project_name = 'Proyecto_Ivo'
 
-data_file = {'tell_classifier':'data_MOTOR-LIBRE.csv',
-            'MCI_classifier':'features_data.csv',
-            'Proyecto_Ivo':'data_total.csv'}
+parallel = True
 
-tasks = {'tell_classifier':['MOTOR-LIBRE'],
-         'MCI_classifier':['fas','animales','fas__animales','grandmean'],
-         'Proyecto_Ivo':['Animales','P','Animales__P','cog','brain','AAL','conn']}
+predefined_models = True
 
-single_dimensions = {'tell_classifier':['voice-quality',
-                                        'talking-intervals','pitch'
-                                        ],
-                     'MCI_classifier':['talking-intervals','psycholinguistic'],
-                     'Proyecto_Ivo':{'Animales':['properties','timing','properties__vr','timing__vr','properties__timing__vr'],
-                                     'P':['properties','timing','properties__vr','timing__vr','properties__timing__vr'],
-                                     'Animales__P': ['properties','timing','properties__vr','timing__vr','properties__timing__vr'],
-                                     'cog':['neuropsico','neuropsico_mmse'],
-                                     'brain':['brain_lit_norm'],
-                                     'AAL':['AAL_norm'],
-                                     'conn':['connectivity']
-                                     }}
-
-test_size = {'tell_classifier':0.3,
-             'MCI_classifier':0.3,
-            'Proyecto_Ivo':0}
-
-dimensions = list()
-if isinstance(single_dimensions[project_name],list):
-    for ndim in range(1,len(single_dimensions[project_name])+1):
-        for dimension in itertools.combinations(single_dimensions[project_name],ndim):
-            dimensions.append('__'.join(dimension))
+l2ocv = False
 
 stratify = False
 
@@ -81,6 +48,46 @@ y_labels = ['target']
 
 id_col = 'id'
 
+cmatrix = None
+shuffle_labels = True
+held_out_default = False
+hyp_tuning_list = [True]
+metrics_names = ['roc_auc','accuracy','recall','f1','norm_expected_cost','norm_cross_entropy']
+
+n_seeds_train = 10
+
+random_seeds_train = np.arange(n_seeds_train) if n_seeds_train > 0 else ['']
+
+thresholds = [0.5]
+
+test_size = {'tell_classifier':0.3,
+             'MCI_classifier':0.3,
+            'Proyecto_Ivo':0}
+
+n_seeds_test_ = 0 if test_size[project_name] == 0 else 1
+
+##---------------------------------PARAMETERS---------------------------------##
+
+data_file = {'tell_classifier':'data_MOTOR-LIBRE.csv',
+            'MCI_classifier':'features_data.csv',
+            'Proyecto_Ivo':'data_total.csv'}
+
+tasks = {'tell_classifier':['MOTOR-LIBRE'],
+         'MCI_classifier':['fas','animales','fas__animales','grandmean'],
+         'Proyecto_Ivo':['Animales','cog','brain','AAL','conn']}
+
+single_dimensions = {'tell_classifier':['voice-quality','talking-intervals','pitch'],
+                     'MCI_classifier':['talking-intervals','psycholinguistic'],
+                     'Proyecto_Ivo':{'Animales':['properties','properties__vr'],
+                                     #'P':['properties','timing','properties_timing','properties__vr','timing__vr','properties__timing__vr'],
+                                     #'Animales__P': ['properties','timing','properties_timing','properties__vr','timing__vr','properties__timing__vr'],
+                                     'cog':['neuropsico','neuropsico_mmse'],
+                                     'brain':['norm_brain_lit'],
+                                     'AAL':['norm_AAL'],
+                                     'conn':['connectivity']
+                                     }}
+
+
 if scaler_name == 'StandardScaler':
     scaler = StandardScaler
 elif scaler_name == 'MinMaxScaler':
@@ -89,24 +96,11 @@ else:
     scaler = None
 imputer = KNNImputer
 
-cmatrix = None
-shuffle_labels = False
-held_out_default = False
-hyp_tuning_list = [True]
-metrics_names = ['roc_auc','accuracy','recall','f1','norm_expected_cost','norm_cross_entropy']
-
-n_seeds_test_ = 0 if test_size[project_name] == 0 else 1
-n_seeds_train = 10
-
 if l2ocv:
     kfold_folder = 'l2ocv'
 else:
     n_folds = 5
     kfold_folder = f'{n_folds}_folds'
-
-random_seeds_train = np.arange(n_seeds_train) if n_seeds_train > 0 else ['']
-
-thresholds = [0.5]
 
 models_dict = {
     'lr':LR,
@@ -119,6 +113,12 @@ data_dir = Path(Path.home(),'data',project_name) if 'Users/gp' in str(Path.home(
 results_dir = Path(str(data_dir).replace('data','results'))
 
 for y_label,task in itertools.product(y_labels,tasks[project_name]):
+    dimensions = list()
+    if isinstance(single_dimensions[project_name],list):
+        for ndim in range(1,len(single_dimensions[project_name])+1):
+            for dimension in itertools.combinations(single_dimensions[project_name],ndim):
+                dimensions.append('__'.join(dimension))
+
     if len(dimensions) == 0:
         dimensions = single_dimensions[project_name][task]
     
@@ -131,7 +131,7 @@ for y_label,task in itertools.product(y_labels,tasks[project_name]):
 
         all_features = [col for col in data.columns if any(f'{x}_{y}__' in col for x,y in itertools.product(task.split('__'),dimension.split('__')))]
         
-        data = data[all_features + [y_label,id_col]].dropna(axis=1,how='any')
+        data = data[all_features + [y_label,id_col]]
         
         features = all_features
         
@@ -156,11 +156,12 @@ for y_label,task in itertools.product(y_labels,tasks[project_name]):
 
             CV_type = StratifiedKFold(n_splits=n_folds,shuffle=True) if stratify else KFold(n_splits=n_folds,shuffle=True)
 
-            path_to_save = Path(results_dir,task,dimension,scaler_name,kfold_folder,y_label,'no_hyp_opt','feature_selection')
+            path_to_save = Path(results_dir,task,dimension,scaler_name,kfold_folder,y_label,'no_hyp_opt','feature_selection','shuffle')
 
             path_to_save = Path(str(path_to_save).replace('no_hyp_opt','hyp_opt')) if hyp_tuning else path_to_save
             path_to_save = Path(str(path_to_save).replace('feature_selection','')) if n_iter_features == 0 else path_to_save
-                            
+            path_to_save = Path(str(path_to_save).replace('shuffle','')) if not shuffle_labels else path_to_save
+            
             path_to_save.mkdir(parents=True,exist_ok=True)
 
             config = {'n_iter':n_iter,
@@ -169,74 +170,75 @@ for y_label,task in itertools.product(y_labels,tasks[project_name]):
             'feature_sample_ratio':feature_sample_ratio,
             'cmatrix':str(cmatrix)}
             
-            hyperp = {'lr': pd.DataFrame({'C': 1},index=[0]),
-                                'lda':pd.DataFrame({'solver':'lsqr'},index=[0]),
-                                'knn': pd.DataFrame({'n_neighbors':5},index=[0]),
-                                'svc': pd.DataFrame({'C': 1,
-                                        'gamma': 'scale',
-                                        'kernel':'rbf',
-                                        'probability':True},index=[0]),
-                                'xgb': pd.DataFrame({'n_estimators':100,
-                                        'max_depth':6,
-                                        'learning_rate':0.3
-                                        },index=[0])
-                    }
-            
-            if hyp_tuning:
-                for n in range(n_iter):
-                    new_combination = dict((key,{}) for key in models_dict.keys())
-                    new_combination['lr'] = {'C': np.random.choice([x*10**y for x,y in itertools.product(range(1,10),range(-3, 2))])}
-                    new_combination['svc'] = {'C': np.random.choice([x*10**y for x,y in itertools.product(range(1,10),range(-3, 2))]),
-                                            'kernel': np.random.choice(['linear', 'rbf', 'sigmoid']),
-                                            'gamma': np.random.choice([x*10**y for x,y in itertools.product(range(1,10),range(-3, 2))]),
-                                            'probability': True}
-                    new_combination['knn'] = {'n_neighbors': int(randint(1, int((n_folds - 1) / n_folds * (data.shape[0] * (1-test_size[project_name])))).rvs())}
-                    new_combination['xgb'] = {'n_estimators': int(randint(10,1000).rvs()),
-                                            'max_depth': randint(1, 10).rvs(),
-                                            'learning_rate': np.random.choice([x*10**y for x,y in itertools.product(range(1,10),range(-3, 2))])
-                                            }
-                    
-                    for key in models_dict.keys():
-                        hyperp[key].loc[len(hyperp[key].index),:] = new_combination[key]
-            
-                hyperp[model].drop_duplicates(inplace=True)
-                hyperp[model] = hyperp[model].reset_index(drop=True)
-            if model == 'knn':
-                hyperp[model] = hyperp[model].astype(int)
-            elif model == 'xgb':
-                hyperp[model] = hyperp[model].astype({'n_estimators':int,'max_depth':int})
+            if predefined_models == False:
+                hyperp = {'lr': pd.DataFrame({'C': 1},index=[0]),
+                                    'lda':pd.DataFrame({'solver':'lsqr'},index=[0]),
+                                    'knn': pd.DataFrame({'n_neighbors':5},index=[0]),
+                                    'svc': pd.DataFrame({'C': 1,
+                                            'gamma': 'scale',
+                                            'kernel':'rbf',
+                                            'probability':True},index=[0]),
+                                    'xgb': pd.DataFrame({'n_estimators':100,
+                                            'max_depth':6,
+                                            'learning_rate':0.3
+                                            },index=[0])
+                        }
+                
+                if hyp_tuning:
+                    for n in range(n_iter):
+                        new_combination = dict((key,{}) for key in models_dict.keys())
+                        new_combination['lr'] = {'C': np.random.choice([x*10**y for x,y in itertools.product(range(1,10),range(-3, 2))])}
+                        new_combination['svc'] = {'C': np.random.choice([x*10**y for x,y in itertools.product(range(1,10),range(-3, 2))]),
+                                                'kernel': np.random.choice(['linear', 'rbf', 'sigmoid']),
+                                                'gamma': np.random.choice([x*10**y for x,y in itertools.product(range(1,10),range(-3, 2))]),
+                                                'probability': True}
+                        new_combination['knn'] = {'n_neighbors': int(randint(1, int((n_folds - 1) / n_folds * (data.shape[0] * (1-test_size[project_name])))).rvs())}
+                        new_combination['xgb'] = {'n_estimators': int(randint(10,1000).rvs()),
+                                                'max_depth': randint(1, 10).rvs(),
+                                                'learning_rate': np.random.choice([x*10**y for x,y in itertools.product(range(1,10),range(-3, 2))])
+                                                }
+                        
+                        for key in models_dict.keys():
+                            hyperp[key].loc[len(hyperp[key].index),:] = new_combination[key]
+                
+                    hyperp[model].drop_duplicates(inplace=True)
+                    hyperp[model] = hyperp[model].reset_index(drop=True)
+                if model == 'knn':
+                    hyperp[model] = hyperp[model].astype(int)
+                elif model == 'xgb':
+                    hyperp[model] = hyperp[model].astype({'n_estimators':int,'max_depth':int})
 
-            num_comb = 0
+                num_comb = 0
 
-            for k in range(np.min((int(feature_sample_ratio*data.shape[0]*(1-test_size[project_name]))-1,len(features)-1))):
-                num_comb += math.comb(len(features),k+1)
+                for k in range(np.min((int(feature_sample_ratio*data.shape[0]*(1-test_size[project_name]))-1,len(features)-1))):
+                    num_comb += math.comb(len(features),k+1)
 
-            feature_sets = list()
+                feature_sets = list()
 
-            if n_iter_features > num_comb:
-                # Generate combinations of features with different lengths
-                for k in range(np.min((int(feature_sample_ratio * data.shape[0] * (1 - test_size)) - 1, len(features) - 1))):
-                    for comb in itertools.combinations(features, k + 1):
-                        feature_sets.append(list(comb))
-            else:
-                # Generate random feature samples
-                feature_sets = [
-                    list(np.unique(np.random.choice(features, int(feature_sample_ratio * data.shape[0] * (1 - test_size)), replace=True)))
-                    for _ in range(n_iter_features)
-                ]
+                if n_iter_features > num_comb:
+                    # Generate combinations of features with different lengths
+                    for k in range(np.min((int(feature_sample_ratio * data.shape[0] * (1 - test_size)) - 1, len(features) - 1))):
+                        for comb in itertools.combinations(features, k + 1):
+                            feature_sets.append(list(comb))
+                else:
+                    # Generate random feature samples
+                    feature_sets = [
+                        list(np.unique(np.random.choice(features, int(feature_sample_ratio * data.shape[0] * (1 - test_size[project_name])), replace=True)))
+                        for _ in range(n_iter_features)
+                    ]
 
-            # Add the full set of features
-            feature_sets.append(features)
+                # Add the full set of features
+                feature_sets.append(features)
 
-            # Remove duplicates by converting to set and back to list of lists
-            feature_sets = [list(t) for t in set(tuple(sorted(fs)) for fs in feature_sets)]
+                # Remove duplicates by converting to set and back to list of lists
+                feature_sets = [list(t) for t in set(tuple(sorted(fs)) for fs in feature_sets)]
 
             for random_seed_test in random_seeds_test:
                                                                                                                                                                                                                                                                                                                                 
-                if test_size > 0:
+                if test_size[project_name] > 0:
                     path_to_save_final = Path(path_to_save,f'random_seed_{random_seed_test}')
 
-                    X_train,X_test,y_train,y_test,ID_train,ID_test = train_test_split(data,y,ID,test_size=test_size,random_state=random_seed_test,shuffle=True,stratify=y if stratify else None)
+                    X_train,X_test,y_train,y_test,ID_train,ID_test = train_test_split(data,y,ID,test_size=test_size[project_name],random_state=random_seed_test,shuffle=True,stratify=y if stratify else None)
                     X_train.reset_index(drop=True,inplace=True)
                     X_test.reset_index(drop=True,inplace=True)
                     y_train.reset_index(drop=True,inplace=True)
@@ -254,8 +256,29 @@ for y_label,task in itertools.product(y_labels,tasks[project_name]):
                     path_to_save_final = path_to_save
 
                 path_to_save_final.mkdir(parents=True,exist_ok=True)
+
+                if predefined_models:
+                    models = pd.read_csv(Path(str(path_to_save_final).replace('shuffle',''),f'all_models_{model}.csv'))
+                    
+                    all_features = models[[col for col in models.columns if any(f'{x}_{y}__' in col for x,y in itertools.product(task.split('__'),dimension.split('__')))]].drop_duplicates()
+
+                    feature_sets = [list([col for col in all_features.columns if all_features.loc[r,col] == 1]) for r in all_features.index]
+                    # Remove duplicates by converting to set and back to list of lists
+                    feature_sets = [list(t) for t in set(tuple(sorted(fs)) for fs in feature_sets)]
+                    hyperp = {model:models[[col for col in models.columns if all(x not in col for x in ['Unnamed: 0','threshold'] + list(all_features.columns))]]}
+                    hyperp[model] = hyperp[model].drop_duplicates()
+                    if 'gamma' in hyperp[model].columns:
+                        for r in hyperp[model].index:
+                            try:
+                                hyperp[model].loc[r,'gamma'] = float(hyperp[model].loc[r,'gamma'])
+                            except:
+                                pass
+
                 assert not set(ID_train).intersection(set(ID_test)), "Data leakeage detected between train and test sets!"
 
+                if Path(path_to_save_final,f'all_models_{model}.csv').exists():
+                    continue
+                
                 with open(Path(path_to_save_final,'config.json'),'w') as f:
                     json.dump(config,f)
